Motor Piezo
There is a program which allows a user to choose a arbitrary
position(between 0 a 75 volts)  for piezo motor and another which sends
a triangular wave to piezo control.
The utility of a triangular wave is the fact that works with constant
velocity, thus its possible to use this program for calibration.