Automation and processing of measurement in optical tweezers.
A control system, processing, and analysis of data to an optic tweezers system. It will be implemented using the software LabVIEW. Basic modules, as auto-focalization, spatial calibration, background image removal, calibration of an optic trap, among others, will be integrated into a control program to a generic user. The development of this research will allow access to modern techniques of organic cells characterization and optic micro-manipulation, saving labor work, energy consumption and the execution time of the activity.